Move to a new build infastructure
Now rather than building everything with pconfigure I have an external
build script and store the results.  This is both good and bad:
specifically, it's got the problem of checking in binaries, but it
should actually work everywhere.

I'm OK with the binaries because this repo doesn't actually do
anything...